Changes to url_list. Dataframe is cleaned. Be sure to read the note.
Url_list -> np.array(1) to [].       -> the reason is that there is a 1 in the url_list which causes a problem when running beautiful soup. 
Data is cleaned. A note: the processed dataframes are in the dictionary teams_dict. The dataframes outside are not changed. 
Year  -> year-month-date to year
W/L -> W -> 1, L-> 0